Final/requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.2
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
boto3==1.34.153
botocore==1.34.153
Brotli==1.1.0
cachetools==5.4.0
calamari-ocr==1.0.7
catalogue==2.0.10
cattrs==23.2.3
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
coremltools==7.2
cryptography==43.0.0
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.3.8
doctr==1.9.0
easyocr==1.7.1
edit-distance==1.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
evaluate==0.4.2
executing==2.0.1
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-vision==3.7.4
google-pasta==0.2.0
googleapis-common-protos==1.63.2
googletrans==4.0.0rc1
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.9.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.8.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.24.3
hyperframe==5.2.0
idna==2.10
imageio==2.34.2
imgaug==0.4.0
importlib_metadata==8.2.0
ipykernel==6.29.5
ipython==8.26.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.1
kiwisolver==1.4.5
kraken==2.0.8
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
lmdb==1.5.1
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mmocr==1.0.1
mplcursors==0.5.3
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.1
pandas==2.2.2
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
preshed==3.0.9
prettytable==3.10.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==24.7.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.18.0
pyparsing==3.1.2
pypdfium2==4.30.0
pyphen==0.16.0
pytesseract==0.3.10
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-doctr==0.7.0
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.5
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scapy==2.5.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.37.0
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
tesseract==0.1.3
tf2onnx==1.16.1
tf_keras==2.17.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.7.24
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
triton==3.0.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
wasabi==1.1.3
watchdog==4.0.1
wcwidth==0.2.13
weasel==0.4.1
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
zopfli==0.2.3
